requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-widget-tweaks==1.5.0
exceptiongroup==1.1.3
facebook-sdk==3.1.0
fastapi==0.104.1
h11==0.14.0
idna==3.4
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pypng==0.20220715.0
python3-openid==3.2.0
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
